modified and cleaned up navigation
